ASC-127 Moved data, exceptions, models, and utils files to a core folder for clearer folder structure
