+++ atlast_sc/parameters/user_input_parameters.py
@@ -1,4 +1,4 @@
-from atlast_sc.utils import Decorators
+from atlast_sc.core.utils import Decorators
 
 ###################################################
 # Getters and setters for user input parameters   #
