+++ atlast_sc_tests/unit_tests/test_models.py
@@ -2,10 +2,10 @@ import math
 
 import pytest
 from astropy import units as u
-from atlast_sc.models import ModelUtils
-from atlast_sc.models import ValueWithUnits, ValueWithoutUnits, \
+from atlast_sc.core.models import ModelUtils
+from atlast_sc.core.models import ValueWithUnits, ValueWithoutUnits, \
     UserInput, CalculationInput
-from atlast_sc.data import Validator
+from atlast_sc.core.data import Validator
 from atlast_sc_tests.utils import does_not_raise
