+++ web_client/calculator.py
@@ -3,7 +3,7 @@ import math
 from dataclasses import asdict
 from atlast_sc.calculator import Calculator
 from pydantic import ValidationError
-from atlast_sc.data import Data
+from atlast_sc.core.data import Data
 
 def do_calculation(user_input, calculation):
     """
